refactor: extract makeSession test helper, dogfood serialize() in integration
/simplify pass on M3+M4 surfaced two findings:

1. **Important**: `makeSession` was triplicated across three new test
   files (`tests/unit/wrapper.test.ts`, `tests/unit/tinyexec-adapter.test.ts`,
   `tests/error-path/tinyexec-errors.test.ts`) with the same body and the
   same comment about the wrapper's lazy-load invariant. Three concrete
   callers at the moment of creation — rule of three triggers cleanly.
   Extracted to `tests/helpers/session.ts`. Net -34 lines; the lazy-load
   invariant comment now lives in one place. Also added a back-reference
   to issue #26 in the helper's comment so the type-system fix is
   linked.

2. **Minor**: `tests/integration/wrapper-tinyexec.test.ts` hand-rolled
   cassette JSON via `JSON.stringify({...})` in two tests. Replaced with
   `serialize(...)` from `src/serialize.js` so the integration tests
   dogfood the canonical-key-order and trailing-newline behavior of
   the real serializer instead of an ad-hoc near-equivalent.

Confirmed clean (no fix): toLines divergence (genuine API difference),
vi.mock setup (vitest hoisting constraint), `OptionsLike` alias
(over-engineering), `version: 1` literal (TS literal type already
enforces). Efficiency review found zero regressions.

Out-of-scope finding flagged for backlog: tinyexec's `aborted: true`
is silently lost on replay (`src/tinyexec.ts:75-77` hardcodes false
because cassette `Result` type has no `aborted` field). Same shape
as the lossy `signal` mapping; tracking together.

All 179 tests still green; lint and typecheck clean.
