+++ tests/unit/tinyexec-adapter.test.ts
@@ -1,8 +1,7 @@
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
-import { DEFAULT_CONFIG } from '../../src/config.js'
-import { MatcherState } from '../../src/matcher.js'
 import { _resetForTesting, clearActiveCassette, setActiveCassette } from '../../src/state.js'
-import type { CassetteSession, Recording } from '../../src/types.js'
+import type { Recording } from '../../src/types.js'
+import { makeSession } from '../helpers/session.js'
 
 vi.mock('tinyexec', () => ({
   x: vi.fn(),
@@ -10,24 +9,6 @@ vi.mock('tinyexec', () => ({
 
 const { x: realXMock } = await import('tinyexec')
 const { x } = await import('../../src/tinyexec.js')
-
-const makeSession = (overrides: Partial<CassetteSession> = {}): CassetteSession => {
-  const base: CassetteSession = {
-    name: 'test',
-    path: '/tmp/test.json',
-    scopeDefault: 'auto',
-    loadedFile: { version: 1, recordings: [] },
-    matcher: null,
-    newRecordings: [],
-    ...overrides,
-  }
-  // Mirror wrapper.ts lazy-load invariant: if loadedFile is set, matcher must
-  // also be set. (Wrapper only inits matcher on the lazy-load path.)
-  if (base.loadedFile !== null && base.matcher === null) {
-    base.matcher = new MatcherState(base.loadedFile.recordings, DEFAULT_CONFIG.matcher)
-  }
-  return base
-}
 
 describe('tinyexec adapter', () => {
   beforeEach(() => {
